requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
gunicorn
django-widget-tweaks
python-magic
pyTelegramBotAPI